src/requirements.txt
actionlib==1.12.1
angles==1.9.12
backports.functools-lru-cache==1.6.4
base-local-planner==1.16.7
bondpy==1.8.5
camera-calibration==1.15.2
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.5.2
certifi==2020.6.20
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cv-bridge==1.13.1
cycler==0.10.0
defusedxml==0.7.1
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
docutils==0.18.1
dynamic-reconfigure==1.6.5
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
image-geometry==1.13.1
interactive-markers==1.11.5
joint-state-publisher==1.12.15
kdl-parser-py==1.13.3
kiwisolver==1.1.0
laser-geometry==1.6.7
matplotlib==2.2.5
message-filters==1.14.13
moveit-core==1.0.11
moveit-ros-planning-interface==1.0.11
netifaces==0.11.0
numpy==1.16.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2023.3
PyYAML==5.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
quaternions==0.6
resource-retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospkg==1.5.0
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.3
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.29
sensor-msgs==1.12.8
six==1.16.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
srdfdom==0.5.2
subprocess32==3.5.4
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.13
urdfdom-py==0.4.6
xacro==1.13.19
